+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [UNRELEASED]
 
 - added support for extension-provided skills
+- using baseCommit when rebasing the worktree on different branch
 
 ## [0.63.0]